% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clip_probs}
\alias{clip_probs}
\title{Clip propensity and probability estimates away from 0 and 1}
\usage{
clip_probs(p, eps = 1e-06)
}
\arguments{
\item{p}{Numeric vector of probabilities.}

\item{eps}{Small positive number (default 1e-6).}
}
\value{
Clipped vector.
}
\description{
Clip propensity and probability estimates away from 0 and 1
}
