% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bounds.R
\name{coef_general_lower_ate}
\alias{coef_general_lower_ate}
\title{Coefficient vectors b such that bound = mean(b' phi)}
\usage{
coef_general_lower_ate()
}
\description{
theta = (E(mu0), E(mu1), E(pi0), E(pi1), E(mu0*pi0), E(mu1*pi1)) = (m1_0, m1_1, m2_0, m2_1, m3_0, m3_1)
Order of phi columns: phi_1_0, phi_1_1, phi_2_0, phi_2_1, phi_3_0, phi_3_1
}
\details{
All bounds/estimands are linear: value = b %*% theta_hat = mean(b %*% t(phi)) = mean(phi %*% b).
}
