% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%||\%}
\alias{\%||\%}
\title{Default value when argument is NULL}
\usage{
x \%||\% y
}
\arguments{
\item{x}{First argument}

\item{y}{Default value to use if x is NULL}
}
\description{
Default value when argument is NULL
}
\keyword{internal}
