% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influence.R
\name{phi_var}
\alias{phi_var}
\title{Empirical variance matrix of phi (for asymptotic variance of b' theta)}
\usage{
phi_var(phi)
}
\arguments{
\item{phi}{n x 6 matrix.}
}
\value{
6 x 6 variance matrix.
}
\description{
Empirical variance matrix of phi (for asymptotic variance of b' theta)
}
