% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influence.R
\name{theta_hat}
\alias{theta_hat}
\title{Compute empirical mean of influence functions (point estimates of theta components)}
\usage{
theta_hat(phi)
}
\arguments{
\item{phi}{n x 6 matrix from influence_functions().}
}
\value{
Vector of length 6: (m1_0, m1_1, m2_0, m2_1, m3_0, m3_1).
}
\description{
Compute empirical mean of influence functions (point estimates of theta components)
}
